+++ library/core/src/test/java/com/google/android/exoplayer2/audio/DefaultAudioSinkTest.java
@@ -15,6 +15,7 @@
  */
 package com.google.android.exoplayer2.audio;
 
+import static com.google.android.exoplayer2.audio.AudioCapabilities.DEFAULT_AUDIO_CAPABILITIES;
 import static com.google.android.exoplayer2.audio.AudioSink.CURRENT_POSITION_NOT_SET;
 import static com.google.android.exoplayer2.audio.AudioSink.SINK_FORMAT_SUPPORTED_DIRECTLY;
 import static com.google.android.exoplayer2.audio.AudioSink.SINK_FORMAT_SUPPORTED_WITH_TRANSCODING;
@@ -24,6 +25,7 @@ import androidx.test.ext.junit.runners.AndroidJUnit4;
 import com.google.android.exoplayer2.C;
 import com.google.android.exoplayer2.Format;
 import com.google.android.exoplayer2.PlaybackParameters;
+import com.google.android.exoplayer2.audio.DefaultAudioSink.DefaultAudioProcessorChain;
 import com.google.android.exoplayer2.util.MimeTypes;
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
@@ -37,7 +39,6 @@ import org.robolectric.annotation.Config;
 /** Unit tests for {@link DefaultAudioSink}. */
 @RunWith(AndroidJUnit4.class)
 public final class DefaultAudioSinkTest {
-
   private static final int CHANNEL_COUNT_MONO = 1;
   private static final int CHANNEL_COUNT_STEREO = 2;
   private static final int BYTES_PER_FRAME_16_BIT = 2;
@@ -59,19 +60,20 @@ public final class DefaultAudioSinkTest {
     arrayAudioBufferSink = new ArrayAudioBufferSink();
     TeeAudioProcessor teeAudioProcessor = new TeeAudioProcessor(arrayAudioBufferSink);
     defaultAudioSink =
-        new DefaultAudioSink(
-            AudioCapabilities.DEFAULT_AUDIO_CAPABILITIES,
-            new DefaultAudioSink.DefaultAudioProcessorChain(teeAudioProcessor),
-            /* enableFloatOutput= */ false,
-            /* enableAudioTrackPlaybackParams= */ false,
-            DefaultAudioSink.OFFLOAD_MODE_DISABLED);
+        new DefaultAudioSink.Builder()
+            .setAudioCapabilities(DEFAULT_AUDIO_CAPABILITIES)
+            .setAudioProcessorChain(new DefaultAudioProcessorChain(teeAudioProcessor))
+            .setOffloadMode(DefaultAudioSink.OFFLOAD_MODE_DISABLED)
+            .build();
   }
 
   @Test
   public void handlesSpecializedAudioProcessorArray() {
     defaultAudioSink =
-        new DefaultAudioSink(
-            AudioCapabilities.DEFAULT_AUDIO_CAPABILITIES, new TeeAudioProcessor[0]);
+        new DefaultAudioSink.Builder()
+            .setAudioCapabilities(DEFAULT_AUDIO_CAPABILITIES)
+            .setAudioProcessors(new TeeAudioProcessor[0])
+            .build();
   }
 
   @Test
@@ -203,10 +205,7 @@ public final class DefaultAudioSinkTest {
   @Test
   public void floatPcmNeedsTranscodingIfFloatOutputDisabled() {
     defaultAudioSink =
-        new DefaultAudioSink(
-            AudioCapabilities.DEFAULT_AUDIO_CAPABILITIES,
-            new AudioProcessor[0],
-            /* enableFloatOutput= */ false);
+        new DefaultAudioSink.Builder().setAudioCapabilities(DEFAULT_AUDIO_CAPABILITIES).build();
     Format floatFormat =
         STEREO_44_1_FORMAT
             .buildUpon()
@@ -221,10 +220,10 @@ public final class DefaultAudioSinkTest {
   @Test
   public void floatPcmNeedsTranscodingIfFloatOutputEnabledBeforeApi21() {
     defaultAudioSink =
-        new DefaultAudioSink(
-            AudioCapabilities.DEFAULT_AUDIO_CAPABILITIES,
-            new AudioProcessor[0],
-            /* enableFloatOutput= */ true);
+        new DefaultAudioSink.Builder()
+            .setAudioCapabilities(DEFAULT_AUDIO_CAPABILITIES)
+            .setEnableFloatOutput(true)
+            .build();
     Format floatFormat =
         STEREO_44_1_FORMAT
             .buildUpon()
@@ -239,10 +238,10 @@ public final class DefaultAudioSinkTest {
   @Test
   public void floatOutputSupportedIfFloatOutputEnabledFromApi21() {
     defaultAudioSink =
-        new DefaultAudioSink(
-            AudioCapabilities.DEFAULT_AUDIO_CAPABILITIES,
-            new AudioProcessor[0],
-            /* enableFloatOutput= */ true);
+        new DefaultAudioSink.Builder()
+            .setAudioCapabilities(DEFAULT_AUDIO_CAPABILITIES)
+            .setEnableFloatOutput(true)
+            .build();
     Format floatFormat =
         STEREO_44_1_FORMAT
             .buildUpon()
@@ -267,8 +266,9 @@ public final class DefaultAudioSinkTest {
   @Test
   public void audioSinkWithAacAudioCapabilitiesWithoutOffload_doesNotSupportAac() {
     DefaultAudioSink defaultAudioSink =
-        new DefaultAudioSink(
-            new AudioCapabilities(new int[] {C.ENCODING_AAC_LC}, 2), new AudioProcessor[0]);
+        new DefaultAudioSink.Builder()
+            .setAudioCapabilities(new AudioCapabilities(new int[] {C.ENCODING_AAC_LC}, 2))
+            .build();
     Format aacLcFormat =
         STEREO_44_1_FORMAT
             .buildUpon()